Move the pane components into own package
